requirements.txt
pyserial==3.5
colorama==0.4.6
pyyaml==6.0.1
requests==2.31.0
argparse